Security Misconfiguration/requirement.txt
Flask
mysql-connector-python
bcrypt
python-dotenv
